requirements.txt
fastmcp>=1.0.0
stellar-sdk>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
